Revert "refactor(workout): extract workout session logic into hooks and helpe…"
This reverts commit 838c56d688854f5bc9b0a5c3689b3b0946ea1e1e.
